% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionial_generator.R
\name{cGAN_generator}
\alias{cGAN_generator}
\title{Title make generator with condition}
\usage{
cGAN_generator(
  node = c(128, 256, 512),
  output_shape = 800L,
  scRNA = NULL,
  custom_count = TRUE,
  count = NULL,
  n.label = 3
)
}
\arguments{
\item{node}{list of nodes in each hidden layer}

\item{output_shape}{Number of cells in modeling}

\item{scRNA}{data for raw or normalized scRNA-seq data,
the ordering of all genes needs to be the same as bulk RNA-seq}

\item{custom_count}{Whether to customize the total number of each single cell (TURE by default)
if set to False, calculate total count of each single cell form normal_single_cell after gene filtering}

\item{count}{total count of each single cell before gene filtering}

\item{n.label}{Number of groups in RNA-seq}
}
\value{
conditionial generator network
}
\description{
Title make generator with condition
}
\examples{
\dontrun{
library(bulkGAN)

generator <- cGAN_generator(node=c(128,256,512),output_shape= 800L,
                          scRNA=normal_single_cell,custom_count=TRUE,count=count,n.label =3)
generator
}
}
