% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Title  make classification for triplet-cGAN}
\usage{
classification(node = c(512, 256, 128, 64, 10), input_shape = 15400L)
}
\arguments{
\item{node}{list of nodes in each hidden layer}

\item{input_shape}{Integer.Number of genes in modeling}
}
\value{
classification network
}
\description{
Title  make classification for triplet-cGAN
}
\examples{
\dontrun{
library(keras)
library(reticulate)
library(tensorflow)
library(bulkGAN)
tfa <- import('tensorflow_addons')  #   import tensorflow_addons model
discriminator <- SNGAN_discriminator(node=c(256,128,64,16,8),
                                     input_shape= 15400L,dropout_rate = 0.2)
discriminator
}
}
