% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_weight.R
\name{get_cell_weight}
\alias{get_cell_weight}
\title{Title get cell weight with h5 file}
\usage{
get_cell_weight(node = c(128, 256, 512), output_shape = 800L, file.path = NULL)
}
\arguments{
\item{node}{list of nodes in each hidden layer}

\item{output_shape}{Number of cells in modeling}

\item{file.path}{File path of file by training for GAN}
}
\value{
generator network for generator cell weight
}
\description{
Title get cell weight with h5 file
}
\examples{
\dontrun{
library(keras)
library(reticulate)
library(tensorflow)
library(bulkGAN)
generator <- SNGAN_generator(node=c(128,256,512),output_shape= 800L,
                            file.path = "~/GAN/2023_2_16/getx/cosin/generator_cosin_1000.h5")

generator

}
