% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNGAN_discriminator.R,
%   R/pairwise_discriminator.R
\name{SNGAN_discriminator}
\alias{SNGAN_discriminator}
\title{Title  make discriminator for SNGAN}
\usage{
SNGAN_discriminator(
  node = c(256, 128, 64, 16, 8),
  input_shape = 15400L,
  dropout_rate = 0.2
)

SNGAN_discriminator(
  node = c(256, 128, 64, 16, 8),
  input_shape = 15400L,
  dropout_rate = 0.2
)
}
\arguments{
\item{node}{list of nodes in each hidden layer}

\item{input_shape}{Integer.Number of genes in modeling}

\item{dropout_rate}{float between 0 and 1. Fraction of the input units to drop.}
}
\value{
discriminator network

discriminator network
}
\description{
Title  make discriminator for SNGAN

Title  make discriminator for pairwise-GAN
}
\examples{
library(keras)
library(reticulate)
library(tensorflow)
library(bulkGAN)
tfa <- reticulate::import('tensorflow_addons')  #   import tensorflow_addons model
discriminator <- SNGAN_discriminator(node=c(256,128,64,16,8),
                                     input_shape= 15400L,dropout_rate = 0.2)
discriminator

library(keras)
library(reticulate)
library(tensorflow)
library(bulkGAN)
tfa <- import('tensorflow_addons')  #   import tensorflow_addons model
discriminator <- SNGAN_discriminator(node=c(256,128,64,16,8),
                                     input_shape= 15400L,dropout_rate = 0.2)
discriminator



}
