% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{SNGAN_generator}
\alias{SNGAN_generator}
\title{Title make generator}
\usage{
SNGAN_generator(
  node = c(128, 256, 512),
  output_shape = 800L,
  scRNA = NULL,
  custom_count = TRUE,
  count = NULL
)
}
\arguments{
\item{node}{list of nodes in each hidden layer}

\item{output_shape}{Number of cells in modeling}

\item{scRNA}{data for raw or normalized scRNA-seq data,
the ordering of all genes needs to be the same as bulk RNA-seq}

\item{custom_count}{Whether to customize the total number of each single cell (TURE by default)
if set to False, calculate total count of each single cell form normal_single_cell after gene filtering}

\item{count}{total count of each single cell before gene filtering}
}
\value{
generator network
}
\description{
Title make generator
}
\examples{
\dontrun{
library(keras)
library(reticulate)
library(tensorflow)
library(bulkGAN)

generator <- SNGAN_generator(node=c(128,256,512),output_shape= 800L,
                           scRNA=NULL,custom_count=TRUE,count=NULL )
generator
}
}
